backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis & Celery
redis==5.0.1
celery==5.3.6

# AWS S3
boto3==1.34.14

# Media Processing
ffmpeg-python==0.2.0

# WhisperX (ASR) - 语音转文字
whisperx==3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# Pyannote.audio - 说话人分离
pyannote.audio>=3.1.0

# LLM - DeepSeek
openai==1.8.0

# Utils
python-ulid==2.2.0
httpx==0.26.0
numpy>=1.24.0
